backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.1
click-log==0.4.0
colorama==0.4.6
cryptography==46.0.6
deprecation==2.1.0
distro==1.9.0
dnspython==2.8.0
docutils==0.22.4
dotty-dict==1.3.1
email-validator==2.3.0
fastapi==0.135.3
fastapi-cli==0.0.24
fastapi-cloud-cli==0.15.1
fastar==0.9.0
gitdb==4.0.12
GitPython==3.1.46
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.1
google-genai==1.70.0
google-generativeai==0.8.6
googleapis-common-protos==1.73.1
gotrue==1.3.1
grpcio==1.80.0
grpcio-status==1.71.2
h11==0.14.0
httpcore==0.16.3
httplib2==0.31.2
httptools==0.7.1
httpx==0.23.3
idna==3.11
importlib_metadata==9.0.0
invoke==1.7.3
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.1.2
jaraco.functools==4.4.0
Jinja2==3.1.6
keyring==25.7.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
more-itertools==11.0.1
nh3==0.3.4
packaging==26.0
pillow==12.2.0
pkginfo==1.12.1.2
postgrest==0.10.7
proto-plus==1.27.2
protobuf==5.29.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-extra-types==2.11.1
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.20.0
pyparsing==3.3.2
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-gitlab==3.15.0
python-multipart==0.0.22
python-semantic-release==7.33.2
pywin32-ctypes==0.2.3
PyYAML==6.0.3
readme_renderer==44.0
realtime==1.0.6
requests==2.33.1
requests-toolbelt==1.0.0
rfc3986==1.5.0
rich==14.3.3
rich-toolkit==0.19.7
rignore==0.7.6
semver==2.13.0
sentry-sdk==2.57.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
starlette==1.0.0
storage3==0.5.3
StrEnum==0.4.15
supabase==1.0.3
supafunc==0.2.2
tenacity==9.1.4
tomlkit==0.14.0
tqdm==4.67.3
twine==3.8.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.42.0
watchfiles==1.1.1
websockets==12.0
wheel==0.46.3
zipp==3.23.0
